Migrate easy q(u)int/bits stuff to torch/headeronly
ghstack-source-id: 2a30e178fd3675eb8409d380baf85ecd25375a26
Pull Request resolved: https://github.com/pytorch/pytorch/pull/159302
